UBUNTU: SAUCE: [power-consumption] allow for fixed compute for cpu, vm, io tests
This patch introduces stress-ng from git://kernel.ubuntu.com/cking/stress-ng.git
rather than using stress.  This allows us to keep the amount of compute
fixed for the cpu, vm and io stress tests so we can therefore end up
measuring both current drawn and also variability in the duration of the
test. Hence we should be able to spot performance improvements or regressions
in the scheduler, cpu scheduler, vm subsystem and the block and ext4 file
system layers.

I've rigorously tested these on identical kit as in the QA lab and made exact
comparisons between the old stress tool and my clean room re-write of stress-ng
to ensure there is little variability between results. stress-ng used in this
updated patch will drive the machine a little harder, so we shalls see some minor
changes in the QA results but from now on we are no longer dependant on variability
that comes from changes in the stress tool now that we are using our own version
that is embedded in this test.

I have also increased the AC current sample rate to 2Hz up from 5Hz to
try to increase the resolution of the sampling and increase the samples
gather by a factor of 2.5 to help improve the statistics.

Signed-off-by: Colin Ian King <colin.king@canonical.com>