UBUNTU: SAUCE: [power_consumption] update with latest instrument-lib tarball
Include new instrument-lib tarball which includes a fix to handle
network connection issues over telnet when connecting to the fluke meter.

This is just some simple exception handling and a few retries before
bailing out.  Errors are logged so we can see the nature of the network
failure rather than a messy python trace dump.

This was tested in the lab by Paul Larsom:

"I'm testing it now, so far so good. I've been through 1 pass this
morning without seeing the previous errors, and I'm running another. I
think between what you've tested and what I've seen so far though,
you're fine to go ahead and push it to Brad."

Signed-off-by: Colin Ian King <colin.king@canonical.com>