UBUNTU: SAUCE: [ubuntu_btrfs_kernel_fixes] Write progress to kernel log
When hags occur it is useful to be able to consule /var/log/syslog and
see which test was invoked that caused the hang later on. So write
the test name to /dev/kmsg

Also, run the tests in alphabetical order rather than some random order
as determined by the file system. This at least allows the tests to be
run in a predictable order.

Signed-off-by: Colin Ian King <colin.king@canonical.com>