UBUNTU: SAUCE: [ubuntu_btrfs_kernel_fixes] remove a test from blacklist
Test for commit 361c093d7f99c3f6cbb07d5c580ce778ab418c42 was failing
so I blacklisted it. After further investigation, all kernels we test
have this fix, and so the failure was not kernel related. It turns out
that I allocated 2G for the btrfs loopback mounted image which was
causing my VM to run low on memory and the OOM killer was nuking
the bash script causing it to fail.  I've dropped the loopback down to
1.4GB and removed the test from the blacklist as it now works fine.

Signed-off-by: Colin Ian King <colin.king@canonical.com>