UBUNTU: SAUCE: [ubuntu_zfs] Add zfs-tests to autotest-client-tests
This commit contains the autotest wrapper for the upstream zfs-tests.
zfs-tests contained in this patch is a snapshot from:

   https://github.com/zfsonlinux/zfs-test.git
   at commit 757261871821ec23eaf69f9fc41971f4981b7cc3
     ("Update the README.md to explain autogen.sh and configure.")

The wrapper pulls in the -stable zfs Ubuntu packages from:
    ppa:zfs-native/stable

..which is based on the zfs-on-linux tarballs from:
    http://zfsonlinux.org/

These tests run to completion, but currently contain a bunch of
failures and killed tests that timeout which need some attention
(to be done later).

The output at the end of the test will be a summary of test results
directly from the zfs-tests on the form:

	Results Summary
	FAIL     147
	SKIP     307
	KILLED    13
	PASS     223

For now, we need to monitor these stats to see how they vary from
kernel to kernel.  Hopefully once I have determined why we see
so many failures and killed tests we can get this into a better state.
The tests contain a 600 second killer that terminates tests if they
are deemed to overrun (showing as "KILLED" tests results).

The zfs-tests run on all Ubuntu kernels from 2.6.32 through to 4.0.0.

Signed-off-by: Colin Ian King <colin.king@canonical.com>