UBUNTU: SAUCE: [ubuntu_zfs_xfs_generic] Add zfs generic xfs tests
This commit contains the autotest wrapper for the upstream xfs generic
tests

xfstests contained in this patch is a snapshot from:

   git://oss.sgi.com/xfs/cmds/xfstests.git
   at commit 78bbab93d8e8a90a2ed0d6cb036ddf36487afa08
     ("config: Fix sysfs paths for partitions")

The wrapper pulls in the -stable zfs Ubuntu packages from:
	ppa:zfs-native/stable

..which is based on the zfs-on-linux tarballs from:
	http://zfsonlinux.org/

This test also patches the xfstests to add functionality to allow
us to run the tests on zfs.  The zfs filesystems are based on a pool
from 2 x 2 mirrored loopback images with 1 journal and 1 cache.

Signed-off-by: Colin Ian King <colin.king@canonical.com>