UBUNTU: SAUCE: [ubuntu_zfs] use a ubuntu specific run script
We need to disable tests that are not currently working or
are known to always fail.  So add a ubuntu specific run script
rather than use the default.

This is currently based on the linux variant but has a couple of
tests disabled that fail because ZFS-on-Linux now working well
with nested ZFS pools in ZFS pools.

Signed-off-by: Colin Ian King <colin.king@canonical.com>