UBUNTU: SAUCE: ubuntu_kernel_selftests: update for new releases
This patch adds control files for missing ubuntu releases. It also
disables the timer test as this randomly fails on RTC wakeup tests
when running inside VMs.  My plan is to figure out why these fail
in the near future, but for now, disable it so we can get at least
some of the selftests back into production again.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>