UBUNTU: SAUCE: ubuntu_zfs_fstest: don't put VDEVs on /tmp and zero VDEV at end
Putting ZFS VDEVs into /tmp can be problematic when a machine is force
rebooted as ZFS tries to mount these and they have been purged, causing ZFS
to wait indefinitely for the VDEVs.

Also, so that there is no trace of the ZFS VDEV data on disk, we completely
overwrite the VDEV with zeros. This may work around some blkid issues where
ZFS is found on the raw device and we don't want this on a reboot.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>