UBUNTU: SAUCE: ubuntu_stress_smoke_test: disable stack stressor
The stack stressor is breaking some user space processes (systemd
journalling and other associated processes) when we consume all
available pages with rapid stack expansion.  I suspect this is
possibly because the OOM mechanism has a different process eviction
mechanism now which maybe causing issues on some architectures.

For now, disable this test as it is rendering systemd et al in
a broken state.

Signed-off-by: Colin Ian King <colin.king@canonical.com>