UBUNTU: SAUCE: ubuntu_loop_smoke_test: new test to test loopback devices
This is a simple test that creates and then destroys 1024 loopback
devices via the /dev/loop-control interface as used by tool such
as losetup.  The test also populates the lookback backing file with
magic patterns that are read back via the /dev/loopN device associated
with the backing store to sanity check the mapping.

Test takes ~7 minutes to run on a Xeon box.

Signed-off-by: Colin Ian King <colin.king@canonical.com>