UBUNTU: SAUCE: ubuntu_kernel_selftests: disable step_after_suspend_test test
The step_after_suspend_test performs a suspend/resume, however, if the wake
alarm time is too long some test machines drop their ssh connections to the
host.  If the alarm is too short we end up not waking up the machine from
suspend.  This is a really fragile test that is causing some breakage on
different test environments, so for now, disable it.

Signed-off-by: Colin Ian King <colin.king@canonical.com>