UBUNTU: SAUCE: lttng smoke test: fix failure counts and context switch counting
I forgot to bump the failure count on some of the test failures, so
add these in. Also make the context switch counting threshold for dd and cat
processeses lower as it does not need to be high, we just need to check
that a handful actually get traced, and this was failing on some slower
virt i386 instances.

Signed-off-by: Colin Ian King <colin.king@canonical.com>