ubuntu_lxc: Fix failure when no http proxy is used
When tests are run outside the ubuntu-ci infrastructure the
http_proxy variable is unset. Testing this variable later will
cause the script to terminate due to the use of "set -u". Change
these tests to use parameter expansion to substitute the string
"none" if the variable is unset or null, and check that the
result != "none" to test whether a proxy is being used.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>