Use the seccomp kselftest instead of ubuntu_seccomp for yakkety/zesty
According to Kees Cook the tests used by ubuntu_seccomp are
deprecated in favor of the seccomp tests in kselftest. The
ubuntu_seccomp tests are failing for y/z anyway due to a bug
in one of the tests, so blacklist ubuntu_seccomp for these
releases. Enable the seccomp kselftests in these releases, and
also in x where they are also present.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>