UBUNTU: SAUCE: ubuntu_kernel_selftests: disable rtctest (LP: #1659333)
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1659333

apw asked me to disable this test: "rtctest interacts badly with
an older xenial HOST kernel in openstack".  This patch sed's out the
rtctest from the Makefile to disable it. Bit of a hack, but it's the
easiest way to do this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>