rtctest: Devices which do not support alarms/interrupts return EINVAL
For some time now devices which do not support alarms or
interrupts will return EINVAL and not ENOTTY in response to
relevant unsupported ioctls. This looks to be true at least as
far back as 3.2.

Update rtctest to expect EINVAL instead of ENOTTY. This
corresponds to changes made to the test in the upstream kernel.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>