UBUNTU: SAUCE: blktrace: add more checking, skip test if blkdev not found
The method of finding the block device is not particularly robust, so
make it find the mount point using stat and then try and find the block
device. If we can't find this, don't abort as a test failure per se, but
instead skip the test instead to stop false positive test failures from
being reported.

Also add a check for blktrace tool - it needs to be installed so double
check this before running the full test.

Signed-off-by: Colin Ian King <colin.king@canonical.com>