UBUNTU: SAUCE: ubuntu_sysdig_smoke_test: make test more solid (LP: #1674936)
On some fast systems there isn't enough time to capture all the
events and so this test fails.  This patch increases the number
off dd blocks being generated, and runs it for at least 4 seconds
to generate plenty of events for sysdig to capture.  We also run
this upto 10 times in case sysdig misses the events.  All we really
want is to ensure that sysdig can run without crashing the kernel
and that it has the capability of capturing events.  Hopefully
this fix will ensure something gets captured(!).

Thanks for Po-Hsu Lin for checking this original issue out and
suggesting upping the threshold.

Signed-off-by: Colin Ian King <colin.king@canonical.com>