ubuntu_fan_smoke_test: Several proxy related fixes
First there was a completely bogues redirection in the -O argument
for wget.
Second, the internal echo command of dash does not know about -e
but bash would require it. So test for the sh link and if it is
dash, do not use -e.
Third, at least for the arm64 node tested the first enablement
proxy is not working with docker. But the second one is. Unfortunately
wget seems to have no issues with both. So there is no way to test.
For now switch probing for those around and hope for the best.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>