UBUNTU: SAUCE: ubuntu_stress_smoke_test: tweak oom setting
We're seeing instances of the autotest script or the driving shell
script getting OOM'd on some of the more aggressive memory hogging
stressors.  To try and fix this, use the --oomable setting on stress-ng
that won't respawn stressors so aggressively if they get oom'd. Also
try and set the OOM level for the driving shell script as high as
possible so it's a less of a candidate to be OOM'd when memory gets
seriously low.

Signed-off-by: Colin Ian King <colin.king@canonical.com>