ubuntu_cts_kernel: Don't suppress perf output for lp1153769
This test has failed a few times with 4.10, but the logs contain
no information about why it failed and I haven't been able to
reproduce the failures locally. The output from perf for the
command is not overwhelming even on success, so stop redireting
it to /dev/null so that failures have something useful about why
the test failed.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>