UBUNTU: SAUCE: ubuntu_blktrace_smoke_test: make test more robust
We've seen this fail on some machines because the test is a bit
flakey.  I've tried to make it more robust by making it perform
far more I/O and pushed the event threshold lower. The change also
includes some smarter waiting and a lot more diagnostics so one
can see timing failures if things go wrong in the future.

Signed-off-by: Colin Ian King <colin.king@canonical.com>