kernel_selftests: Fix breakpoints setup and re-enable
The makefile in the Linux kernel tree changed the naming of the
variable which stores the sub-commands to run. This broke the
script that tried to disable one sub-command.
Replace the pattern matching with a more generic version tha works
with all names of the variable.

This also re-enables the breakpoints selftest again on artful.

BugLink: https://bugs.launchpad.net/bugs/1706283

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>