UBUNTU: SAUCE: force blktrace test to run on a device that causes trace events (LP: #1712254)
When running on some devices (such as virtualized s390x) the underlying device
does not generate block trace events, so this test fails. Overcome this by
running the test on a loopback device with ext4 on top; this always generate
block device events hence is always traceable on all architectures.

Signed-off-by: Colin Ian King <colin.king@canonical.com>