UBUNTU: SAUCE: ubuntu_sysdig_smoke_test: limit sysdig capture file (LP: #1714190)
On fast systems it is easy for sysdig to generate millions of trace
events that generate a massive sysdig capture file. This can fill up the
file system causing the system to clog up and not progress because
we're out of free space.  Fix this by limiting the raw trace file to 16MB
which is plenty to capture all the events we need.

This fix also termimates faster when sysdig terminates early. Also added
is an improved killing of a locked up sysdig and some minor cleanups
on variable name references.

Signed-off-by: Colin Ian King <colin.king@canonical.com>