UBUNTU: SAUCE: wait for sysdig to load and start (LP: #1717884)
On some devices sysdig takes a while to get loaded and start up, so
add a wait delay that waits for sysdig to emit a "starting capture"
kernel message before we start gathering sysdig events. Also add a
debug message to dump the amount of time waited and another debug
message to report if the capture message was not emitted in time.

Signed-off-by: Colin Ian King <colin.king@canonical.com>