UBUNTU: SAUCE: Add new ZRAM smoke test
This test exercises the kernel ZRAM compress ramdisk.  In this
smoke test we load the module, create a ZRAM backed ext4 file
system, populated it with random and zero content files and
also exercise some of the ZRAM /sys interfaces.  If the kernel
supports different compression algorithms we try to exercise
all of these too.

Tested on Ubuntu Precise, Trusty, Xenial, Artful and Bionic.

Signed-off-by: Colin Ian King <colin.king@canonical.com>