UBUNTU: SAUCE: ubuntu_zfs_xfs_generic: bump xfsprogs version
Use the latest xfsprogs release to fix the compilation issue with glibc 2.27
BugLink: https://launchpad.net/bugs/1753987

Patch has been sent to the upstream xfstest-bld repo, we can revert this
once the merge request got accepted.

I didn't put this change into the 0002 patch for config file since in this
way, it will be easier to revert this commit in the future.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>