kernel_selftests: fix disabling of step_after_suspend_test
On latest kernels, including bionic, a change on the breakpoints selftest
Makefile caused our disabling it to not work. So, we were running this test on
ppc64el systems, and that broke them. The Makefile pattern changed from += to
:=. Accepting any optional character before equals still works for artful as
well, when the previous pattern was present, and let's hope it won't break
again too soon.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>