UBUNTU: SAUCE: ubuntu_zfs_xfs_geneirc: Add support for arm64 (LP: #1755499)
It turns out that the reason why xattr xfs test fails on ZFS is because
the xattr tests don't have the xattr syscalls wired up to any actual
syscalls on arm6 and are hard-coded to always return ENOSYS.  Fix this
by adding in the arm64 syscall support for the xattr family of syscalls.

Also move and comment out the zfs umount $POOL as I believe it is not required
but I've left it in just in case we need to enable it in the future and moved
it to where I expect it could be more useful to use.

Just want to add that figuring out this fix was a pain as I had to work
through the ZFS + vfs kernel stack back down to userspace until I discovered
that the xfs tests bypass glibc and user their own xattr compat syscall shims
which was not what I expected.

Signed-off-by: Colin Ian King <colin.king@canonical.com>