UBUNTU SAUCE: kernel-smoke-test-scrub: allow for resilvering to complete
On systems where I/O is slow, resilvering may take a few seconds so
add a 60 second resilvering loop check to give it time to finish
before forcing a scrub for the scrub test.  This fixes a race in the
ADT testing as reported by Stefan

Also don't dd using /dev/urandom to create a large data file as
reading from /dev/urandom is slow. Instead generate several K of
random data and write this multiple times using 'yes' as this is
far faster and speeds up the test on slower CPU machines.

Signed-off-by: Colin Ian King <colin.king@canonical.com>