UBUNTU: SAUCE: ubuntu_kvm_smoke_test: fix the ssh key permission issue
As we're running this test as root, it will try to use the key in root's home
directory instead of the one we created for the user, and the uvt-kvm wait
command will fail with:
    Permission denied (publickey).

This only happens when testing with the remote script in kernel-testing.

Explictly define the ssh key file to use for the uvt-kvm wait command and the
ssh command to fix this problem.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>