UBUNTU: SAUCE: ubuntu_docker_smoke_test: request for a image explicitly
The docker run command was not used with a specific release name tag,
it will go fetch the image with the 'latest' tag automatically.

After Bionic was released, this latest was changed from Xenial to Bionic.
To avoid version check failure in the future, this patch will ask it to
run a certain release explicitly.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>