UBUNTU:SAUCE: ubuntu_ltp_syscalls: seprate the syscalls test from the main ltp test suite.
Pull the syscalls test out of ubuntu_ltp test suite to replace the old syscall
fuzzer test scrashme.

Using this test to replace scrashme is because Trinity, the successor of
scrashme, will took a huge amount of time and storage space to do fuzz testing
on syscalls. Which is not ideal for regression-testing. And the report
generated by LTP is more readable.

Not running this syscalls test in the ubuntu_ltp test suite so that we can gain
more control over which release to test, which kind of instances to test.

Tested on P(amd64), T(amd64), X(arm64, s390x), A(s390x), B (arm64, KVM), it
cannot be build on Precise thus blacklisted on it.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>