UBUNTU SAUCE: ubuntu_docker_smoke_test: skip arch check for docker image
BugLink: https://bugs.launchpad.net/bugs/1775574

The arch check code with ARM64 nodes will compose: aarch64/ubuntu:bionic
which is not correct for the project hosted on docker hub:
    docker: Error response from daemon: manifest for aarch64/ubuntu:bionic
    not found.

It should be arm64v8/ubuntu:bionic instead.

Now the docker application is smart enough to fetch the image for the
corresponding arch, there is no need to run this check anymore.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>