UBUNTU SAUCE: ubuntu_docker_smoke_test: fix Trusty i386 docker test
BugLink: https://bugs.launchpad.net/bugs/1779109

The arch check was skipped in commit 0af3172.
However the docker application in Trusty is not smart enough yet to get the
corresponding arch with just ubuntu:bionic.

Let's partially revert the i386 part for that commit.

docker.io package for ARM64/ppc is not available on Trusty, so we can just add
the i386 check.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>