UBUNTU SAUCE: ubuntu_bpf: move setup steps to the python script
Move the test setup steps from the shell script to the python script.
So that we could group those steps together and make the test itself looks tidy.

The git repo cleanup can be ignored after this, as it will be cloned into the
src directory, not the result directory.

It also paves the road to split setup task from the test itself in the future.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>