UBUNTU: SAUCE: ubuntu_zram_smoke_test: try hard to avoid OOMing
Instead of using a relatively large zram size, make one that is
one third the size of the free available memory size. Ensure we
also have a zram size of least 8MB to ensure a useable ext4 file
system can ber exercised on this zram device.

This should avoid any out of memory OOMing on very low memory
instances. I've tested this with an instance of 768MB and it works
fine with this fix.

Signed-off-by: Colin Ian King <colin.king@canonical.com>