UBUNTU: SAUCE: ubuntu_ramfs_smoke_test: try hard to avoid OOMing
Instead of using a relatively large ram size, make one that is
scaled appropriately to the memory available.

Also clean up the stress-ng tests to exercise just file system
based testing and remove some stressors that were accidentally
left in and contribute to memory bloat rather than file I/O
activity.

Finally, remove VFAT_IMAGE0 as this is not required and was
left over from the original test that this one was copied from.

This should avoid any out of memory OOMing on very low memory
instances. I've tested this with an instance of 768MB and it works
fine with this fix.

Signed-off-by: Colin Ian King <colin.king@canonical.com>