UBUNTU SAUCE: split the setup task for huge tests
Split the setup task for those tests that need to be build from scratch.
This makes the report become shorter for parser / tester to read.

Also, it will make it easier to identify if an issue happens on the setup
stage or the test itself has failed.

The message length (lines) for setup tasks:
  * 18000 - ubuntu_ltp / ubuntu_ltp_syscalls
  * 6300 - xfstests
  * 1100 - ubuntu_blktrace_smoke_test
  * 1100 - ubuntu_kvm_unit_tests
  * 900 - ubuntu_kvm_smoke_test
  * 900 - ubuntu_ramfs_stress
  * 900 - ubuntu_vfat_stress
  * 850 - ubuntu_stress_smoke_test
  * 500 - ubuntu_kernel_selftests
  * 400 - ubuntu_loop_smoke_test
  * 200 - ubuntu_bpf
  * 100 - ubuntu_lxc
  * 100 - ubuntu_zfs_smoke_test

For ubuntu_lttng_smoke_test, the cleanup task will be triggered right after
the setup task, so it can't be changed like this.

Patch tested on our jenkins server, the result looks ok.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>