UBUNTU: SAUCE: install the extra module required by the test
BugLink: https://bugs.launchpad.net/bug/1777114

As we have the linux-modules-extra meta package for the GCP and Azure cloud
now, we can add it to the package list for ubuntu_quota_smoke_test and
ubuntu_zram_smoke_test, they both need the module in the extra package to run.

Use the asterisk here to solve the package detection for azure and azure-edge.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>