UBUNTU SAUCE: ubuntu_vfat_stress: workaround hangs on slow backing device (LP: #1786462)
BugLink: https://bugs.launchpad.net/bug/1786462

We're getting lockups on VMs with slow backing devices with this
vfat stressor. I've managed to reproduce this issue with various
kernels by running the test with a ZFS backing store to the VM
with sync writes enabled for each write to a slow HDD. Since we
are interesting in smoke testing the vfat file system and not
how the block layer performs when many sync'd I/Os become blocked
on a slow device, I've decided to put the VFS into a loop-backed
mounted file system on top of a tmpfs file system.  I've tested
this on a 1GB memory VM with a very slow backing device and this
still works fine as we're balancing low memory and I/O with swap
instead of sending all I/Os through to a slow file based vfat
image.

Also, bump each vfat mount option to run for 10 seconds as this
allows all the stress tests to exercise the vfat file systems
more thoroughly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>