UBUNTU SAUCE: ubuntu_zram_smoke_test: reduce dd size, add timeout cleanup (LP: #1787076)
BugLink: https://bugs.launchpad.net/bug/1787076

On comparitively low compute systems with large memory the size of
the file being dd'd onto the zram device is taking a long time to
compress and the test is being nuke'd on a timeout.  Add a upper
and lower file size limit so that the test won't take a long time
to run even on slow CPUs, and hence won't timeout.  The test is now
taking ~5 minutes even on a mcdivitt box, which is way below the
20 minute timeout.

Also add a umount of the mounted zram file system if we interrupt
the test to try to clean up. Note this won't work for timeout
terminations as autotest kills the test before we can reap resouces.

Signed-off-by: Colin Ian King <colin.king@canonical.com>