ubuntu_cts_kernel/lp1262692: use dummy net device instead of veth pair
This test is using an odd syntax to set up a veth device for the
test, passing ip the default device. In practice it looks like
this was just getting ignored in the past, however starting with
iproute2 4.18.0-1ubuntu2 this syntax is being rejected.

The test looks to only be creating a veth pair to have a dummy
interface for the tests. A dummy device makes more sense here, so
switch to using that, which has the side effect of also fixing
the ip command syntax issue.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>