UBUNTU: SAUCE: ubuntu_ramfs_smoke_test: don't create huge files
Specify the hdd stressor filesize to fit reasonably into the
ramfs file system. Currently the default is 256MB which can cause
kernel OOMing, so reduce it to something that will be small enough
to stress the ramfs but not cause OOMing of the autotest or test
environment.

Also bump the test duration to 30 seconds to ensure the stressor
exercises the ramfs more thoroughly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>