UBUNTU SAUCE: ubuntu_sysdig_smoke_test: skip for T, run for X-4.15
As stated in the initial commit of this test from Colin:
    sysdig landed in Zesty and is a rather useful tracing tool

This test should be working for kernel newer than Zesty and onward.

After testing on X-4.4 / X-4.15 and T-4.15 Azure and discussed with Colin,
we decided not to run this on T-4.15 Azure as the tool chain is a bit too
old for this (bug 1796016).

This test can pass with X-4.15, so change the Xenial series specific blacklist
to become 4.4.0 kernel specific.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>