UBUNTU SAUCE: Add new stress-ng performance based benchmarking tests
This adds a new performance benchmarking based on various stress-ng
tests.  The tests selected have been shown to have low variation in
bogo op throughput measurements across a range of hardware types,
kernels, architectures (s390, x86-64 and arm64) and baremetal and
VM configurations.

Each stress test is run 5 times, each run for 60 seconds. The average
and standard deviation of these results is calculated and dumped to
the output for later parsing and adding to a database (such as influxdb)
for posterity and for performance stats trend tracking graphs.

Typical output per test is:

14:26:05 INFO | clock_bogoops 100.134 100.117 100.150 100.133 100.117
14:26:05 INFO | clock_mean 100.130
14:26:05 INFO | clock_stddev 0.014
14:26:05 INFO | clock_percent_stddev 0.014
14:26:05 INFO | clock_expected_threshold 0.040
14:26:05 INFO | clock_within_threshold True

..showing the bogoops measurements, mean, stddev, percentage stddev and
the expected percentage stddev threshold. If the test falls within the
threshold (+10% slop) then <test-name>_within_threshold is True.

Signed-off-by: Colin Ian King <colin.king@canonical.com>