UBUNTU SAUCE: ubuntu_zram_smoke_test: kill multipath on /dev/zram (LP: #1798291)
BugLink: https://bugs.launchpad.net/bugs/1798291

Some systems have multipath that kicks in and grabs zram, so
forcefully kill it so that the test can work without any
interference.

multipath on /dev/zram was causing issues such as not allowing
one to open the associated sysfs files and being able to unload
the driver.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>