UBUNTU SAUCE: ubuntu_performance_pts: disable DynamicRunCount
The Phoronix Test Suite will run a test for more times if the std.dev.
is outside an artificial threshold. Rather than run the test an
unpredictable number of time we should set DynamicRunCount to FALSE
to disable this statistically questionable default.  We have to slam
the config file into /etc and into the home directory to ensure we
pick up the correct one.

Signed-off-by: Colin Ian King <colin.king@canonical.com>