UBUNTU SAUCE: ubuntu_vfat_stress: fix vfat hangs (LP: #1801907)
Running on fast multiple processor systems with a lot of memory
can lead to this test hanging when using the vfat flush mount
option that forces writeback on all writes.  Running stress-ng
that saturates the backing device with writes a slow performing
flush on each write can lead to the flushes becoming backed up
causing timeout warnings. The hangs occur when stress-ng termimates
but the stress child processes are blocked in async writes and
the main stress parent is waiting for the children to eventually
unblock.

I think the flush mount option can be skipped for now as we need
the tests to run in a short time rather than hours waiting for it
to complete the stalled writes.

This commit disables flush and also adds in a polling loop to
monitor stress-ng and forcefully terminate it if it overruns. I've
also added more debug information that is written to /dev/kmsg so
that any future hangs are easier to diagnose

Signed-off-by: Colin Ian King <colin.king@canonical.com>