UBUNTU SAUCE: ubuntu_vfat_stress: use loopback instead of tmpfs
For systems that are low on memory and/or have a lot of I/O stressors
we get low memory contention when mounting a 1GB vfat file system
on a tmpfs backing file. Re-work the test so that we use a simpler
loopback mounted vfat file system with direct I/O also disabled.

This seems to help with issues where we get dirty page write
amplification because of using a very large tmpfs with low memory.

Signed-off-by: Colin Ian King <colin.king@canonical.com>