UBUNTU SAUCE: ubuntu_ltp_syscalls: install extra modules for cloud kernels
BugLink: https://bugs.launchpad.net/bugs/1790028

The quotactl01 test will need the quota module provided in the
linux-extra-modules package to work.

Install it for Azure, Google cloud kernels.

For AWS, quota is not supported as it does not provide the extra module
package (lp:1777813), the test should be disabled for AWS.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>