UBUNTU SAUCE: ubuntu_vfat_stress: ensure stress-ng is killed (LP: #1806359)
Killing stress-ng just once may not be enough, it can be in the process
of re-spawning new stressors causing the umount of the vfat loop
filesystem to fail later. Double check no stress-ng processes are
accessing the mounted vfat filesystem with lsof, and retry the kill
to ensure all existing stress-ng stressors are really dead.

Signed-off-by: Colin Ian King <colin.king@canonical.com>