UBUNTU: SAUCE: nbd: fix race on ndb mounting (LP: #1806013)
BugLink: https://bugs.launchpad.net/bugs/1806013

Using the old-style nbd server configuration wasn't a great idea,
so use a simple auto-generated nbd configuration file that allows
us to also list the nbd exports.  This way the test can probe
to see if the nbd server is running and exporting the file so
that the client and always reliably mount it.  This seems to fix
the occasional nbd client mount failures.

Signed-off-by: Colin Ian King <colin.king@canonical.com>