UBUNTU: SAUCE: nbd: fix race on device becoming usable (LP: #1807869)
It appears that on slow systems (such as ones that may have some
contention) there is a race on the nbd server creating the nbd device
and the point where it becomes usable to format. Make the test more
resiliant by retrying a format if it fails and giving up after 20
iterations.

Signed-off-by: Colin Ian King <colin.king@canonical.com>