UBUNTU SAUCE: Add new power performance benchmark
The purpose of this test is to run this on a specific machine and
to measure the CPU power consumption of the CPU package for several
different CPU stress tests from stress-ng.

Each test scenario is run 5 times with an average/min/max/stddev CPU
power consumption computed from the results with a threshold sanity
check.  Data can be scraped from the results log and pushed into
a suitable database for weekly trend analysis.

Typical output is as follows:

18:11:18 INFO | matrix_all_cpus_x86_watts 38.530 38.590 38.700 38.820 38.940
18:11:18 INFO | matrix_all_cpus_x86_watts_minimum 38.530
18:11:18 INFO | matrix_all_cpus_x86_watts_maximum 38.940
18:11:18 INFO | matrix_all_cpus_x86_watts_average 38.716
18:11:18 INFO | matrix_all_cpus_x86_watts_maximum_error 1.059%
18:11:18 INFO | matrix_all_cpus_x86_watts_stddev 0.167
18:11:18 INFO | matrix_all_cpus_x86_watts_percent_stddev 0.432
18:11:18 INFO | PASS: test passes specified performance thresholds

Signed-off-by: Colin Ian King <colin.king@canonical.com>