UBUNTU SAUCE: ubuntu_kernel_selftests: fix timer failures
The fix for LP: #1659333 no longer works for newer kernels because
the rtc test has been moved to it's own directory, so add a check
to cater for either paths to disable this test.

Secondly, the rtcpie test will always fail on systems that don't
have an x86 style /dev/rtc0 device, so add a check and disable that
test for these types of systems.

Signed-off-by: Colin Ian King <colin.king@canonical.com>