UBUNTU SAUCE: ubuntu_kernel_selftests: fix cpu hotplug
The cpu-hotplug test assumes that we can offline the maximum CPU as
described by /sys/devices/system/cpu/offline.  However, in the case
where the number of CPUs exceeds like kernel configuration then
the offline count can be greater than the present count and we end
up trying to test the offlining of a CPU that is not available to
offline.  Fix this by testing the maximum present CPU instead.

Also, the test currently offlines the CPU and does not online it,
so fix this by onlining the CPU after the test.

Note: It it hard to patch the test because of the different kernel
sources that it needs to be applied to. So I've taken the liberty
to just copy in the fixed script from linux 5.0-rc1 plus my fix that
I sent upstream https://lkml.org/lkml/2019/1/10/291 - the copy also
checks for older named versions of the script too and handles that
case too.  The copy only occurs if the script does not contain my
upstream fix, so it won't occur once my fix lands in later kernels.

Signed-off-by: Colin Ian King <colin.king@canonical.com>