UBUNTU: SAUCE: ubuntu_stress_smoke_test: skip sctp for 4.20
The sctp stressor is known to break on 4.20, so disable it. Since
we will be using 5.0 and not 4.20 for Disco we can be slack here
and skip this test.

Also skip bind-mount as this stressor is skipped by stress-ng anyway
and skip funccall ioport as these are not kernel related tests.

Signed-off-by: Colin Ian King <colin.king@canonical.com>