UBUNTU SAUCE: ubuntu_kernel_selftests: fail the test suite if a sub-test has failed
For the ADT testing, it depends on the return code to decide if a test
suite should be marked as failed or not.

This creates some false-positive situation when a sub test in the suite
has failed, but since the test suite itself has been executed
successfully, it makes ADT mark it as green.

Fix this by using a flag combine with checking the test result for each
sub-test to determine if we should fail this suite in the end.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>