UBUNTU SAUCE: ubuntu_ftrace_smoke_test: fix timeout handling and child process killing
The timeout handling is overly complex using SIGALRM handling; instead
re-work this to just have a child process that sleeps and kills the
test if a timeout is reached. This fixes some asynchronous issues when
older sleeping timeouts trigger when near to the end of the test.

Also fix the child process killing on the reading of the tracing pipe,
the wrong processes were being killed and a pending cat on the pipe
made it subsequently unreadable. Doh.

Also force sync of data to file system to ensure all the traces of
writes occur due to the sync flushing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>