UBUNTU: SAUVE: kvm-unit-tests: use our own repo
kvm-unit-tests follows KVM development very closely, which means its tests tend
to not be backwards compatible. That means we should be running the tests we
know to work on a given kernel version and watch for regressions on the kernel,
not regressions on the tests itself.

Right now, kvm-unit-tests is useless for watching regressions because it has
always failed on lots of tests, and it's really hard to find out what has
regressed and what not.

So, using our own version of kvm-unit-tests makes sense on trying to watch for
real regressions on the KVM side. Right now, there is only a disco branch. What
we really need is a for-v5.0.0 branch instead, that will be fixed in the
future.

As we don't have branches for other kernels yet, test this one for now. At
least, this will allow us to get more code fixed for our different systems once
we get that tested, and start watching for regressions on 5.0 kernels, before
we invest too much on the other kernel versions.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>