UBUNTU SAUCE: ubuntu_ftrace_smoke_test: remove kprobes check
BugLink: https://bugs.launchpad.net/bugs/1821907

The check for kprobes is not actually required for this smoke
test as all we are checking is for basic ftrace functionality
and early ARM64 kernels (xenial and before) don't support
kprobes.  Remove the kprobes check.

Also double the timeout on the function graph tracing test
as this takes a while to complete on slower systems.

Finally, use an array for the sysfs tracing file checks as
this is easier to break into lines for a tidier looking test.

Signed-off-by: Colin Ian King <colin.king@canonical.com>